Minor refactor on audiobook bookmark related classes
